refactor: helper function inside the other file
